+++ meridian/analysis/analyzer.py
@@ -611,18 +611,26 @@ def _scale_tensors_by_multiplier(
     data: DataTensors,
     multiplier: float,
     by_reach: bool,
+    non_media_treatments_baseline: tf.Tensor | None = None,
 ) -> DataTensors:
   """Get scaled tensors for incremental outcome calculation.
 
   Args:
-    data: DataTensors object containing the optional tensors to scale: `media`,
-      `reach`, `frequency`.
+    data: DataTensors object containing the optional tensors to scale. Only
+      `media`, `reach`, `frequency`, `organic_media`, `organic_reach`,
+      `organic_frequency`, `non_media_treatments` are scaled. The other tensors
+      remain unchanged.
     multiplier: Float indicating the factor to scale tensors by.
     by_reach: Boolean indicating whether to scale reach or frequency when rf
       data is available.
+    non_media_treatments_baseline: Optional tensor to overwrite
+      `data.non_media_treatments` in the output. Used to compute the
+      conterfactual values for incremental outcome calculation. If not used, the
+      unmodified `data.non_media_treatments` tensor is returned in the output.
 
   Returns:
-    A `DataTensors` object containing scaled tensor parameters.
+    A `DataTensors` object containing scaled tensor parameters. The original
+    tensors that should not be scaled remain unchanged.
   """
   incremented_data = {}
   if data.media is not None:
@@ -634,6 +642,32 @@ def _scale_tensors_by_multiplier(
     else:
       incremented_data[constants.REACH] = data.reach
       incremented_data[constants.FREQUENCY] = data.frequency * multiplier
+  if data.organic_media is not None:
+    incremented_data[constants.ORGANIC_MEDIA] = data.organic_media * multiplier
+  if data.organic_reach is not None and data.organic_frequency is not None:
+    if by_reach:
+      incremented_data[constants.ORGANIC_REACH] = (
+          data.organic_reach * multiplier
+      )
+      incremented_data[constants.ORGANIC_FREQUENCY] = data.organic_frequency
+    else:
+      incremented_data[constants.ORGANIC_REACH] = data.organic_reach
+      incremented_data[constants.ORGANIC_FREQUENCY] = (
+          data.organic_frequency * multiplier
+      )
+  if non_media_treatments_baseline is not None:
+    incremented_data[constants.NON_MEDIA_TREATMENTS] = (
+        non_media_treatments_baseline
+    )
+  else:
+    incremented_data[constants.NON_MEDIA_TREATMENTS] = data.non_media_treatments
+
+  # Include the original data that does not get scaled.
+  incremented_data[constants.MEDIA_SPEND] = data.media_spend
+  incremented_data[constants.RF_SPEND] = data.rf_spend
+  incremented_data[constants.CONTROLS] = data.controls
+  incremented_data[constants.REVENUE_PER_KPI] = data.revenue_per_kpi
+
   return DataTensors(**incremented_data)
 
 
@@ -1720,6 +1754,7 @@ class Analyzer:
       aggregate_times: bool = True,
       inverse_transform_outcome: bool = True,
       use_kpi: bool = False,
+      by_reach: bool = True,
       include_non_paid_channels: bool = True,
       batch_size: int = constants.DEFAULT_BATCH_SIZE,
   ) -> tf.Tensor:
@@ -1735,10 +1770,11 @@ class Analyzer:
     by `media_selected_times`. Similarly, `Media_0` means that media execution
     is multiplied by `scaling_factor0` (0.0 by default) for these time periods.
 
-    For channels with reach and frequency data, the frequency is held fixed
-    while the reach is scaled. "Outcome" refers to either `revenue` if
-    `use_kpi=False`, or `kpi` if `use_kpi=True`. When `revenue_per_kpi` is not
-    defined, `use_kpi` cannot be False.
+    For channels with reach and frequency data, either reach or frequency is
+    held fixed while the other is scaled, depending on the `by_reach` argument.
+    "Outcome" refers to either `revenue` if `use_kpi=False`, or `kpi` if
+    `use_kpi=True`. When `revenue_per_kpi` is not defined, `use_kpi` cannot be
+    False.
 
     If `new_data=None`, this method computes incremental outcome using `media`,
     `reach`, `frequency`, `organic_media`, `organic_reach`, `organic_frequency`,
@@ -1828,6 +1864,10 @@ class Analyzer:
         otherwise the expected revenue `(kpi * revenue_per_kpi)` is calculated.
         It is required that `use_kpi = True` if `revenue_per_kpi` data is not
         available or if `inverse_transform_outcome = False`.
+      by_reach: Boolean. If `True`, then the incremental outcome is calculated
+        by scaling the reach and holding the frequency constant. If `False`,
+        then the incremental outcome is calculated by scaling the frequency and
+        holding the reach constant. Only used for channels with RF data.
       include_non_paid_channels: Boolean. If `True`, then non-media treatments
         and organic effects are included in the calculation. If `False`, then
         only the paid media and RF effects are included.
@@ -1928,34 +1968,15 @@ class Analyzer:
         ]
     non_media_selected_times = media_selected_times[-mmm.n_times :]
 
-    # Set counterfactual media and reach tensors based on the scaling factors
-    # and the media selected times.
+    # Set counterfactual tensors based on the scaling factors and the media
+    # selected times.
     counterfactual0 = (
         1 + (scaling_factor0 - 1) * np.array(media_selected_times)
     )[:, None]
     counterfactual1 = (
         1 + (scaling_factor1 - 1) * np.array(media_selected_times)
     )[:, None]
-    new_media0 = (
-        None
-        if data_tensors.media is None
-        else data_tensors.media * counterfactual0
-    )
-    new_reach0 = (
-        None
-        if data_tensors.reach is None
-        else data_tensors.reach * counterfactual0
-    )
-    new_organic_media0 = (
-        None
-        if data_tensors.organic_media is None
-        else data_tensors.organic_media * counterfactual0
-    )
-    new_organic_reach0 = (
-        None
-        if data_tensors.organic_reach is None
-        else data_tensors.organic_reach * counterfactual0
-    )
+
     if data_tensors.non_media_treatments is not None:
       new_non_media_treatments0 = _compute_non_media_baseline(
           non_media_treatments=data_tensors.non_media_treatments,
@@ -1964,55 +1985,23 @@ class Analyzer:
       )
     else:
       new_non_media_treatments0 = None
-    new_media1 = (
-        None
-        if data_tensors.media is None
-        else data_tensors.media * counterfactual1
+
+    incremented_data0 = _scale_tensors_by_multiplier(
+        data=data_tensors,
+        multiplier=counterfactual0,
+        by_reach=by_reach,
+        non_media_treatments_baseline=new_non_media_treatments0,
     )
-    new_reach1 = (
-        None
-        if data_tensors.reach is None
-        else data_tensors.reach * counterfactual1
+    incremented_data1 = _scale_tensors_by_multiplier(
+        data=data_tensors, multiplier=counterfactual1, by_reach=by_reach
     )
-    new_organic_media1 = (
-        None
-        if data_tensors.organic_media is None
-        else data_tensors.organic_media * counterfactual1
-    )
-    new_organic_reach1 = (
-        None
-        if data_tensors.organic_reach is None
-        else data_tensors.organic_reach * counterfactual1
-    )
-    new_non_media_treatments1 = (
-        None
-        if data_tensors.non_media_treatments is None
-        else data_tensors.non_media_treatments
-    )
+
     data_tensors0 = self._get_scaled_data_tensors(
-        new_data=DataTensors(
-            media=new_media0,
-            reach=new_reach0,
-            frequency=data_tensors.frequency,
-            organic_media=new_organic_media0,
-            organic_reach=new_organic_reach0,
-            organic_frequency=data_tensors.organic_frequency,
-            non_media_treatments=new_non_media_treatments0,
-            revenue_per_kpi=data_tensors.revenue_per_kpi,
-        ),
+        new_data=incremented_data0,
         include_non_paid_channels=include_non_paid_channels,
     )
     data_tensors1 = self._get_scaled_data_tensors(
-        new_data=DataTensors(
-            media=new_media1,
-            reach=new_reach1,
-            frequency=data_tensors.frequency,
-            organic_media=new_organic_media1,
-            organic_reach=new_organic_reach1,
-            organic_frequency=data_tensors.organic_frequency,
-            non_media_treatments=new_non_media_treatments1,
-            revenue_per_kpi=data_tensors.revenue_per_kpi,
-        ),
+        new_data=incremented_data1,
         include_non_paid_channels=include_non_paid_channels,
     )
 
@@ -2201,14 +2190,6 @@ class Analyzer:
         "selected_times": selected_times,
         "aggregate_geos": aggregate_geos,
     }
-    incremental_outcome_kwargs = {
-        "inverse_transform_outcome": True,
-        "use_posterior": use_posterior,
-        "use_kpi": use_kpi,
-        "batch_size": batch_size,
-        "include_non_paid_channels": False,
-        "aggregate_times": True,
-    }
     self._validate_geo_and_time_granularity(**dim_kwargs)
     required_values = [
         constants.MEDIA,
@@ -2225,20 +2206,19 @@ class Analyzer:
         meridian=self._meridian,
         allow_modified_times=False,
     )
-    incremental_outcome = self.incremental_outcome(
+    numerator = self.incremental_outcome(
         new_data=filled_data,
-        **incremental_outcome_kwargs,
+        scaling_factor0=1,
+        scaling_factor1=1 + incremental_increase,
+        inverse_transform_outcome=True,
+        use_posterior=use_posterior,
+        use_kpi=use_kpi,
+        by_reach=by_reach,
+        batch_size=batch_size,
+        include_non_paid_channels=False,
+        aggregate_times=True,
         **dim_kwargs,
     )
-    incremented_data = _scale_tensors_by_multiplier(
-        data=filled_data,
-        multiplier=incremental_increase + 1,
-        by_reach=by_reach,
-    )
-    incremental_outcome_with_multiplier = self.incremental_outcome(
-        new_data=incremented_data, **dim_kwargs, **incremental_outcome_kwargs
-    )
-    numerator = incremental_outcome_with_multiplier - incremental_outcome
     spend_inc = filled_data.total_spend() * incremental_increase
     if spend_inc is not None and spend_inc.ndim == 3:
       denominator = self.filter_and_aggregate_geos_and_times(
@@ -2874,11 +2854,6 @@ class Analyzer:
         "aggregate_geos": aggregate_geos,
         "aggregate_times": aggregate_times,
     }
-    dim_kwargs_wo_agg_times = {
-        "selected_geos": selected_geos,
-        "selected_times": selected_times,
-        "aggregate_geos": aggregate_geos,
-    }
     batched_kwargs = {"batch_size": batch_size}
     aggregated_impressions = self.get_aggregated_impressions(
         new_data=new_data,
@@ -2911,6 +2886,32 @@ class Analyzer:
         non_media_baseline_values=non_media_baseline_values,
         **dim_kwargs,
         **batched_kwargs,
+    )
+    incremental_outcome_mroi_prior = self.compute_incremental_outcome_aggregate(
+        use_posterior=False,
+        new_data=new_data,
+        use_kpi=use_kpi,
+        by_reach=marginal_roi_by_reach,
+        scaling_factor0=1,
+        scaling_factor1=1 + marginal_roi_incremental_increase,
+        include_non_paid_channels=include_non_paid_channels,
+        non_media_baseline_values=non_media_baseline_values,
+        **dim_kwargs,
+        **batched_kwargs,
+    )
+    incremental_outcome_mroi_posterior = (
+        self.compute_incremental_outcome_aggregate(
+            use_posterior=True,
+            new_data=new_data,
+            use_kpi=use_kpi,
+            by_reach=marginal_roi_by_reach,
+            scaling_factor0=1,
+            scaling_factor1=1 + marginal_roi_incremental_increase,
+            include_non_paid_channels=include_non_paid_channels,
+            non_media_baseline_values=non_media_baseline_values,
+            **dim_kwargs,
+            **batched_kwargs,
+        )
     )
     expected_outcome_prior = self.expected_outcome(
         use_posterior=False,
@@ -3079,19 +3080,14 @@ class Analyzer:
           confidence_level=confidence_level,
           spend_with_total=spend_with_total,
       )
-      mroi = self._compute_marginal_roi_aggregate(
-          marginal_roi_by_reach=marginal_roi_by_reach,
-          marginal_roi_incremental_increase=marginal_roi_incremental_increase,
-          expected_revenue_prior=expected_outcome_prior,
-          expected_revenue_posterior=expected_outcome_posterior,
+      mroi = self._compute_roi_aggregate(
+          incremental_outcome_prior=incremental_outcome_mroi_prior,
+          incremental_outcome_posterior=incremental_outcome_mroi_posterior,
           xr_dims=xr_dims_with_ci_and_distribution,
           xr_coords=xr_coords_with_ci_and_distribution,
           confidence_level=confidence_level,
-          spend_with_total=spend_with_total,
-          new_data=new_data,
-          use_kpi=use_kpi,
-          **dim_kwargs_wo_agg_times,
-          **batched_kwargs,
+          spend_with_total=spend_with_total * marginal_roi_incremental_increase,
+          metric_name=constants.MROI,
           # Drop mROI metric values in the Dataset's data_vars for the
           # aggregated "All Paid Channels" channel dimension value.
           # "Marginal ROI" calculation must arbitrarily assume how the
@@ -4442,87 +4438,13 @@ class Analyzer:
       xr_coords: Mapping[str, tuple[Sequence[str], Sequence[str]]],
       spend_with_total: tf.Tensor,
       confidence_level: float = constants.DEFAULT_CONFIDENCE_LEVEL,
+      metric_name: str = constants.ROI,
   ) -> xr.Dataset:
     # TODO: Support calibration_period_bool.
     return _central_tendency_and_ci_by_prior_and_posterior(
         prior=incremental_outcome_prior / spend_with_total,
         posterior=incremental_outcome_posterior / spend_with_total,
-        metric_name=constants.ROI,
-        xr_dims=xr_dims,
-        xr_coords=xr_coords,
-        confidence_level=confidence_level,
-        include_median=True,
-    )
-
-  def _compute_marginal_roi_aggregate(
-      self,
-      marginal_roi_by_reach: bool,
-      marginal_roi_incremental_increase: float,
-      expected_revenue_prior: tf.Tensor,
-      expected_revenue_posterior: tf.Tensor,
-      xr_dims: Sequence[str],
-      xr_coords: Mapping[str, tuple[Sequence[str], Sequence[str]]],
-      spend_with_total: tf.Tensor,
-      new_data: DataTensors | None = None,
-      use_kpi: bool = False,
-      confidence_level: float = constants.DEFAULT_CONFIDENCE_LEVEL,
-      **roi_kwargs,
-  ) -> xr.Dataset:
-    if new_data is None:
-      new_data = DataTensors()
-    data_tensors = new_data.validate_and_fill_missing_data(
-        [constants.MEDIA, constants.REACH, constants.FREQUENCY], self._meridian
-    )
-    mroi_prior = self.marginal_roi(
-        use_posterior=False,
-        new_data=data_tensors,
-        by_reach=marginal_roi_by_reach,
-        incremental_increase=marginal_roi_incremental_increase,
-        use_kpi=use_kpi,
-        **roi_kwargs,
-    )
-    mroi_posterior = self.marginal_roi(
-        use_posterior=True,
-        new_data=data_tensors,
-        by_reach=marginal_roi_by_reach,
-        incremental_increase=marginal_roi_incremental_increase,
-        use_kpi=use_kpi,
-        **roi_kwargs,
-    )
-    incremented_data = _scale_tensors_by_multiplier(
-        data=data_tensors,
-        multiplier=(1 + marginal_roi_incremental_increase),
-        by_reach=marginal_roi_by_reach,
-    )
-
-    mroi_prior_total = (
-        self.expected_outcome(
-            use_posterior=False,
-            new_data=incremented_data,
-            use_kpi=use_kpi,
-            **roi_kwargs,
-        )
-        - expected_revenue_prior
-    ) / (marginal_roi_incremental_increase * spend_with_total[..., -1])
-    mroi_posterior_total = (
-        self.expected_outcome(
-            use_posterior=True,
-            new_data=incremented_data,
-            use_kpi=use_kpi,
-            **roi_kwargs,
-        )
-        - expected_revenue_posterior
-    ) / (marginal_roi_incremental_increase * spend_with_total[..., -1])
-    mroi_prior_concat = tf.concat(
-        [mroi_prior, mroi_prior_total[..., None]], axis=-1
-    )
-    mroi_posterior_concat = tf.concat(
-        [mroi_posterior, mroi_posterior_total[..., None]], axis=-1
-    )
-    return _central_tendency_and_ci_by_prior_and_posterior(
-        prior=mroi_prior_concat,
-        posterior=mroi_posterior_concat,
-        metric_name=constants.MROI,
+        metric_name=metric_name,
         xr_dims=xr_dims,
         xr_coords=xr_coords,
         confidence_level=confidence_level,